+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+## v0.1.0-rc.3
+
 - Fix conflicts with same method names across different modules on iOS by prefixing method names
 - Update Dart duration serialization to be compatible with Kotlin ISO-8601 duration string parsing